@@ -5,6 +5,7 @@ import CharList from "../charList/CharList";
 import CharInfo from "../charInfo/CharInfo";
 import decoration from '../../resources/img/vision.png';
 import ErrorBoundary from "../errorBoundary/ErrorBoundary";
+import ComicsList from "../comics-list/ComicsList";
 
 
 const  App = () => {
@@ -18,8 +19,9 @@ const  App = () => {
         <div className="app">
             <AppHeader/>
             <main>
-                <RandomChar/>
-                <div className="char__content">
+                {/* <RandomChar/> */}
+                <ComicsList/>
+                {/* <div className="char__content">
                     <CharList 
                     onCharSelected={onCharSelected}
                     charId={selectedChar}/>
@@ -28,7 +30,7 @@ const  App = () => {
                         charId={selectedChar}/>
                     </ErrorBoundary>
                 </div>
-                <img className="bg-decoration" src={decoration} alt="vision"/>
+                <img className="bg-decoration" src={decoration} alt="vision"/> */}
             </main>
         </div>
     )
